Upload of new version of the code
Code updated to publish data on a new topic. The type of data published is a custom one that is defined in the msg folder in .msg file (the c++ header file is generated with the command catkin_make install). Another utility script was added to let generate the header file. Modification of the xml and CMakeList files in order to use custom msg.